AdvancedStore - Joseph Gutierrez - Maria Garcia/advancedStoreFinal/requirements.txt
flask
pymongo
python-dotenv
dnspython